Refactored to make more robust
