+++ src/message.ts
@@ -1,0 +1,6 @@
+import { WorkBook } from "xlsx"
+
+export interface Message {
+    type: 'initialRequestCompleted' | 'spreadsheet' | 'error'
+    data: undefined | WorkBook | any
+}
