+++ src/worker.ts
@@ -1,13 +1,38 @@
+import * as XLSX from "xlsx"
+import { Message } from "./message";
+
+// A service worker will not unload with an open port,
+//   so it is safe to store this port here
+let port: chrome.runtime.Port | undefined = undefined;
+
+chrome.runtime.onConnect.addListener(p => { port = p });
+
 chrome.webRequest.onCompleted.addListener(async ({ url }) => {
     try {
-        const [tab] = await chrome.tabs.query({ active: true, currentWindow: true })
-        chrome.tabs.sendMessage(tab.id!, url)
+        port!.postMessage(<Message>{ type: 'initialRequestCompleted' })
+        const workbook = await fetchWorkdaySpreadsheet(url)
+        port!.postMessage(<Message>{ type: 'spreadsheet', data: workbook })
     } catch (exception) {
-        console.log('Failed to notify the current tab of the request url', exception)
+        console.log('Failed to send the data to the port: ', exception)
+        port?.postMessage(<Message>{ type: 'error', data: exception })
+    } finally {
+        port?.disconnect()
+        port = undefined
     }
 }, { urls: ["https://wd5.myworkday.com/*"] })
 
-// The service worker doesn't wake up like it is supposed to for webRequest
-// See https://bugs.chromium.org/p/chromium/issues/detail?id=1024211
-// This message listener is to force the service worker to wake up when needed
-chrome.runtime.onMessage.addListener(console.log)
+/**
+ * Jumps through hoops to fetch the actual excel file Workday exports
+ * 
+ * @param url the url used by Workday to request the excel spreadsheet
+ * @returns the excel spreadsheet
+ */
+function fetchWorkdaySpreadsheet(url: string) {
+    return fetch(url)
+        .then(response => response.json())
+        .then(json => json.docReadyUri)
+        .then(path => new URL(url).origin + path)
+        .then(resourceLocation => fetch(resourceLocation))
+        .then(response => response.arrayBuffer())
+        .then(bytes => XLSX.read(bytes, { type: 'array' }))
+}
